Revert "Fix LightDM missing package error/warning (Debian #837979)"
Installing `accountsservice` fixes LightDM warnings (Debian #837979).
But it also enables a new Accounts Service.  The benefit of running
this service is debatable, and there is an outstanding bug with Account
Server causing LightDM to freeze and/or leak memory (Debian #908293).
It may be safer to just ignore the warnings and keep using LightDM
without the Accounts Service.

References:
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=837979
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=908293